Provide basic instructions for setting up server
A GitHub user would expect to know how to run this project from reading the README. But the README does not provide these instruction; a beginner has to read about python and Django to know the basic commands. So tell the user what to do to get a running web server.